updated sitecard component and moved to sitelist
